mitty-6vz.13: Quality gate — fix CodeRabbit findings
- Add env var overrides for AI config settings (RPM, TPM, budgets)
- Fix coach endpoint created_at="" → fetch actual DB timestamp
- Fix _escape_like backslash order (escape backslash before % and _)
- Handle null values in ai_usage aggregation
- Fix FK type mismatch in db.py (BigInteger → Integer for study_block_id)
- Add ge=0/ge=1 validation to coach pagination params
- Update tests for new behavior

+++ mitty/db.py
@@ -584,7 +584,7 @@ coach_messages = sa.Table(
     sa.Column("user_id", sa.Uuid, nullable=False),
     sa.Column(
         "study_block_id",
-        sa.BigInteger,
+        sa.Integer,
         sa.ForeignKey("study_blocks.id"),
         nullable=False,
     ),
